Refactored code
Moved ui functions into different files, used translation texts, removed discontinued packages.
