+++ lib/features/core/presentation/providers/is_router_link_active_provider.dart
@@ -7,6 +7,6 @@ part 'is_router_link_active_provider.g.dart';
 @riverpod
 bool isRouterLinkActive(IsRouterLinkActiveRef ref, AppLocations appLocation) {
   final activeRouteNotifier = ref.watch(activeRouteProvider);
-  // ignore: deprecated_member_use
-  return activeRouteNotifier.value.location == appLocation.href;
+
+  return activeRouteNotifier.value.uri.toString() == appLocation.href;
 }
