+++ lib/features/core/presentation/providers/router_notifier.dart
@@ -3,7 +3,7 @@ import 'package:field_companion/features/core/presentation/views/home.dart';
 import 'package:field_companion/features/core/presentation/views/welcome.dart';
 import 'package:field_companion/features/field_service/presentation/views/field_service.dart';
 import 'package:field_companion/features/settings/presentation/providers/app_initialisation_provider.dart';
-import 'package:field_companion/features/settings/presentation/widgets/about_the_app.dart';
+import 'package:field_companion/features/settings/presentation/widgets/about_the_app/about_the_app.dart';
 import 'package:field_companion/features/settings/presentation/widgets/settings.dart';
 import 'package:field_companion/features/territory_cards/presentation/views/territories.dart';
 import 'package:flutter/material.dart';
