+++ lib/features/core/presentation/widgets/common/section_item_styles.dart
@@ -15,9 +15,17 @@ class SectionItemStyles {
 
   static TextStyle whiteKey = _key.copyWith(color: Colors.white);
 
+  static TextStyle blackKey = _key.copyWith(color: Colors.black);
+
   static TextStyle blueKey = _key.copyWith(color: ColorPalette.blue);
 
   static TextStyle redKey = _key.copyWith(color: ColorPalette.red);
 
   static TextStyle value = _key.copyWith(color: ColorPalette.grey2);
+
+  static TextStyle valueBold = _key.copyWith(
+    color: Colors.white,
+    fontSize: 18,
+    fontWeight: FontWeight.bold,
+  );
 }
