+++ pubspec.yaml
@@ -42,9 +42,9 @@ dependencies:
   isar_flutter_libs: ^3.0.5
   logger: ^1.3.0
   nanoid: ^1.0.0
-  package_info: ^2.0.2
+  package_info_plus: ^5.0.1
   riverpod_annotation: ^2.0.4
-  share: ^2.0.4
+  share_plus: ^7.2.2
   shared_preferences: ^2.1.0
   stack_trace: ^1.11.0
